+++ Cargo.toml
@@ -38,5 +38,8 @@ async-process = "1.5"
 serde = "1.0"
 # For parsing info.json
 serde_json = "1.0"
+
 # For parsing language identifiers
-isolang = "2.0"
+[dependencies.isolang]
+features = ["serde"]
+version = "2.0"
